Small fixes/modifications.

+++ Payment.Api/Models/PaymentModel.cs
@@ -8,7 +8,7 @@ namespace Payment.Api.Models
         [Key]
         public int PaymentId { get; set; }
         public DateTime CreationDate { get; set; }
-        public uint Amount { get; set; }
+        public decimal Amount { get; set; }
         public string CurrencyCode { get; set; }
         public string Status { get; set; }
         public OrderModel Order { get; set; }
